KYLIN: Makefile: fix generated/compile.h missing builder
Mainline: KYLIN-only
Severity: Low

Ideally, we would like to fix the compilation user of the kernel to
'KYLINSOFT', but because the previous modification was too late, it was
not correctly imported into compile.h. We submitted this code to the
front, which can effectively fix the builder.

K2CI-Arch: All
Fixes: 80319bbc92f5 ("KYLIN: config: Init x86 and arm64 default configure")
Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: Iae9973f8263b1b0ee0e831054e0bf5542428d3e9
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/101345
Tested-by: k2ci <kernel-bot@kylinos.cn>