HYGON: linux: tcm: add Hygon TCM2 driver
Mainline: KYLIN-only
Severity: Important

task: 216398

Hygon CPU implemented a firmware-based TCM2 device, which runs on its
internal secure processor named PSP. The device underlying uses an unique
private interface in the form of some hardware mailbox between X86 cores
and PSP, which is for sure different from the TIS or CRB interfaces defined
in the PTP spec.

As such, to support this device we need a specialized driver which handles
the basic send and receive operations required by the kernel TPM core
layer. ACPI device info passed from underlying BIOS indicates the device
presence by setting the _HID field to "HYGT0201", which distinguishes
it from the rest of devices. If the BIOS does not support this setting,
the driver will not be activated and thus has no impact to the system
at all.

Signed-off-by: chench <chench@hygon.cn>
Signed-off-by: shitao <shitao@kylinos.cn>
Change-Id: Ib84d092d083f66d44209e00442bcc563b86467b6
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/114323
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>