tools: Add ethtool.h header to tooling infra
Mainline: 7effe3fdc049a34c56a68671100b5570e53e8f0a
From: v6.10-rc1
Severity: Moderate

task: 341439

This commit duplicates the ethtool.h file from the include/uapi/linux
directory in the kernel source to the tools/include/uapi/linux directory.

This action ensures that the ethtool.h file used in the tools directory
is in sync with the kernel's version, maintaining consistency across the
codebase.

There are some checkpatch warnings in this file that could be cleaned up,
but I preferred to move it over as-is for now to avoid disrupting the code.

Signed-off-by: Tushar Vyavahare <tushar.vyavahare@intel.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Acked-by: Magnus Karlsson <magnus.karlsson@intel.com>
Link: https://lore.kernel.org/bpf/20240402114529.545475-2-tushar.vyavahare@intel.com

[Backport notes]
Due to the lack of a series of ethtool_* definitions in selftest/bpf,
this will cause compilation errors, so it is necessary to introduce this
community patch to resolve the compilation issue.

This commit is the first of seven series commits. The reason it hasn't
been merged with the other six is that it would introduce a large number
of additional changes. This commit simply copies the ethtool.h file from
the include/uapi/linux directory in the kernel source to
the tools/include/uapi/linux directory, without affecting functionality,
so I've only merged this one.

K2CI-Arch: All
Signed-off-by: luoxuanqiang <luoxuanqiang@kylinos.cn>
Change-Id: If17d539b77a30615a816ca4160b5e48a4cefa238
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/128906
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>