mlxsw: pci: Arm CQ doorbell regardless of number of completions
Mainline: 38b124cb4ee5d3567cd138af33bbd6fefbfdb703
From: v6.10-rc1
Severity: Low

task: 507933

Currently, as part of mlxsw_pci_cq_tasklet(), we check if any item
was handled, and only in such case we arm doorbell. This is unlikely case,
as we schedule tasklet only for CQs that we get an event for them, which
means that they contain completions to handle. Remove this check, which
is supposed to be true always, and even if it is false, it is not a mistake
to ring the doorbell. We can warn on such case, but it is not really worth
to add a check which will be run for each CQ handling when we do not expect
to reach it and it does not point to logic error that should be handled.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>
Reviewed-by: Ido Schimmel <idosch@nvidia.com>
Signed-off-by: Petr Machata <petrm@nvidia.com>
Reviewed-by: Simon Horman <horms@kernel.org>
Link: https://lore.kernel.org/r/f8efa481bfe7bebb9f93bb803f44ab7da77f53e6.1712062203.git.petrm@nvidia.com
Signed-off-by: Jakub Kicinski <kuba@kernel.org>

(backported from commit 38b124cb4ee5d3567cd138af33bbd6fefbfdb703)

Signed-off-by: Wang Yan <wangyan01@kylinos.cn>
Change-Id: Ia8b4bb307b5992bbab5116c0fa4f827c3611bcc5
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/96223
Tested-by: kernel-bot <kernel-bot@kylinos.cn>
Reviewed-by: 刘云 <liuyun01@kylinos.cn>
Code-Style: kernel-bot <kernel-bot@kylinos.cn>