mlxsw: pci: Break mlxsw_pci_cq_tasklet() into tasklets per queue type
Mainline: 1df7d871e3491de2f53afafedef8383185a3f613
From: v6.10-rc1
Severity: Low

task: 507933

Completion queues are used for completions of RDQ or SDQ. Each
completion queue is used for one DQ. The first CQs are used for SDQs and
the rest are used for RDQs.

Currently, for each CQE (completion queue element), we check 'sr' value
(send/receive) to know if it is completion of RDQ or SDQ. Actually, we
do not really have to check it, as according to the queue number we know
if it handles completions of Rx or Tx.

Break the tasklet into two - one for Rx (RDQ) and one for Tx (SDQ). Then,
setup the appropriate tasklet for each queue as part of queue
initialization. Use 'sr' value for unlikely case that we get completion
with type that we do not expect. Call WARN_ON_ONCE() only after checking
the value, to avoid calling this method for each completion.

A next patch set will use NAPI to handle events, then we will have a
separate poll method for Rx and Tx. This change is a preparation for
NAPI usage.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>
Reviewed-by: Ido Schimmel <idosch@nvidia.com>
Signed-off-by: Petr Machata <petrm@nvidia.com>
Reviewed-by: Simon Horman <horms@kernel.org>
Link: https://lore.kernel.org/r/50fbc366f8de54cb5dc72a7c4f394333ef71f1d0.1712062203.git.petrm@nvidia.com
Signed-off-by: Jakub Kicinski <kuba@kernel.org>

(backported from commit 1df7d871e3491de2f53afafedef8383185a3f613)

Signed-off-by: Wang Yan <wangyan01@kylinos.cn>
Change-Id: I5bd881a8e9bee908a0967a04c9ca701ab280a521
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/96229
Reviewed-by: 刘云 <liuyun01@kylinos.cn>
Code-Style: kernel-bot <kernel-bot@kylinos.cn>
Tested-by: kernel-bot <kernel-bot@kylinos.cn>