crypto: hisilicon - Adjust debugfs creation and release order
Mainline: 0a6e038f0c4d9a5d5f638757fa03ef6ffc550b6a
From: v6.10-rc1
Severity: Moderate

task: 435539

There is a scenario where the file directory is created but the
file memory is not set. In this case, if a user accesses the
file, an error occurs.

So during the creation process of debugfs, memory should be
allocated first before creating the directory. In the release
process, the directory should be deleted first before releasing
the memory to avoid the situation where the memory does not
exist when accessing the directory.

In addition, the directory released by the debugfs is a global
variable. When the debugfs of an accelerator fails to be
initialized, releasing the directory of the global variable
affects the debugfs initialization of other accelerators.
The debugfs root directory released by debugfs init should be a
member of qm, not a global variable.

Signed-off-by: Chenghai Huang <huangchenghai2@huawei.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>

(backported from commit 0a6e038f0c4d9a5d5f638757fa03ef6ffc550b6a)

K2CI-Arch: All
Signed-off-by: zhouyuhang <zhouyuhang@kylinos.cn>
Change-Id: I8381256a73ae8d3830f6328904f47f455647b5bc
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/158801
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>