ZHAOXIN: x86/delay: add support for Zhaoxin ZXPAUSE instruction
Mainline: KYLIN-only
Severity: Important

task: 209974

ZXPAUSE instructs the processor to enter an implementation-dependent
optimized state. The instruction execution wakes up when the time-stamp
counter reaches or exceeds the implicit EDX:EAX 64-bit input value.
The instruction execution also wakes up due to the expiration of
the operating system time-limit or by an external interrupt.

ZXPAUSE is available on processors with X86_FEATURE_ZXPAUSE.
ZXPAUSE allows the processor to enter a light-weight power/performance
optimized state (C0.1 state) for a period specified by the instruction
or until the system time limit.

MSR_ZX_PAUSE_CONTROL MSR register allows the OS to enable/disable C0.2 on
the processor and to set the maximum time the processor can reside in C0.1
or C0.2. By default C0.2 is disabled.

A sysfs interface to adjust the time and the C0.2 enablement is provided in
a follow up change.

Signed-off-by: leoliu-oc <leoliu-oc@zhaoxin.com>
Signed-off-by: shitao <shitao@kylinos.cn>
Change-Id: Ie61f5d14432af6241479e9311c6e621dafe998f2
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/106324
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>