OPENEULER: Revert "scsi: hisi_sas: Disable SATA disk phy for severe I_T nexus reset failure"
Mainline: KYLIN-only
Severity: Moderate

Task: #219034

In that commit, if the softreset fails upon certain conditions, just
disable the PHY associated with the disk. The user needs to restore the
PHY.

SATA disks do not support simultaneous connection of multiple hosts.
Therefore, when multiple controllers are connected to a SATA disk at the
same time, the controller which is connected later failed to issue an ATA
softreset to the SATA disk. As a result, the PHY associated with the disk
is disabled and cannot be automatically recovered.

Now that, we will not focus on the execution result of softreset. No
matter whether the execution is successful or not, we will directly carry
out I_T_nexus_reset.

Signed-off-by: Yihang Li <liyihang9@huawei.com>
Signed-off-by: Bing Xia <xiabing12@h-partners.com>

K2CI-Arch: All
Link: https://gitee.com/openeuler/kernel/issues/I96KNQ
Signed-off-by: Yongzhen Zhang <zhangyongzhen@kylinos.cn>
Change-Id: I9bf8fd485623d59ec7a28f79d7818e8d4d09ace9
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/113115
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>