HYGON: crypto: tdm: Support dynamic protection for SCT and IDT by HYGON TDM
Mainline: KYLIN-only
Severity: Important

task: 216398

tdm_kernel_guard is an application that uses HYGON TDM technology
to protect important data in the kernel. Through this application,
the dynamic protection of SCT and IDT is completed in the system.
In the future, more protection objects can be expanded based on this
application

Signed-off-by: chench <chench@hygon.cn>
Signed-off-by: shitao <shitao@kylinos.cn>
Change-Id: I894e02920b6a48d858df1386f75a3d984ccd1b82
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/113683
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>
Reviewed-by: 孟德慧 <mengdehui@kylinos.cn>
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>