genirq: Provide a snapshot mechanism for interrupt statistics
Mainline: 99cf63c56661be0a0c42f79b56f37a4aa34b4779
From: v6.10-rc1
Severity: Low

task: 350471

The soft lockup detector lacks a mechanism to identify interrupt storms as
root cause of a lockup. To enable this the detector needs a mechanism to
snapshot the interrupt count statistics on a CPU when the detector observes
a potential lockup scenario and compare that against the interrupt count
when it warns about the lockup later on. The number of interrupts in that
period give a hint whether the lockup might have been caused by an interrupt
storm.

Instead of having extra storage in the lockup detector and accessing the
internals of the interrupt descriptor directly, add a snapshot member to
the per CPU irq_desc::kstat_irq structure and provide interfaces to take a
snapshot of all interrupts on the current CPU and to retrieve the delta of
a specific interrupt later on.

Originally-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Bitao Hu <yaoma@linux.alibaba.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Link: https://lore.kernel.org/r/20240411074134.30922-3-yaoma@linux.alibaba.com

(backported from commit 99cf63c56661be0a0c42f79b56f37a4aa34b4779)

Signed-off-by: Jackie Liu <liuyun01@kylinos.cn>
Change-Id: I87b6b17e9da73c089522fbc121e9f1e2e711d443
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/130646
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>