ZHAOXIN: x86/mce/zhaoxin: Enable mcelog to decode PCIE, ZDI/ZPI, and DRAM errors
Mainline: KYLIN-only
Severity: Important

task: 209974

The mcelog cannot decode PCIE, ZDI/ZPI, and DRAM errors in the FFM
(Firmware First Mode).
The purpose of this patch is to enable mcelog to decode PCIE, ZDI/ZPI, and
DRAM errors that occur on Zhaoxin processors, so that the cause of these
errors can be quickly located.

Signed-off-by: leoliu-oc <leoliu-oc@zhaoxin.com>
Signed-off-by: shitao <shitao@kylinos.cn>
Change-Id: I06804c0df7e67910e18ccc575282585a3f62d458
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/106274
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>