wifi: mac80211: mlme: handle cross-link CSA
Mainline: 7ef8f6821d16623a3f6ea279542f9f47ece9ff1e
From: v6.11-rc1
Severity: Moderate

task: 496179

Backport note:
Conflict in:
net/mac80211/mlme.c
1. This patch change sdata_info to link_info.
2. Fix conflict in after rebase.

Origin message:
If we see a channel switch announcement on one link for
another, handle that case and start the CSA. The driver
can react to this in whatever way it needs. The stack
will have the ability to track it via the RNR/MLE in the
reporting link's beacon if it sees it for inactive links
and adjust everything accordingly.

Note that currently the timings for the CSA aren't set,
the values are only used by the Intel drivers, and they
don't need this for newer devices that support MLO, so
I've left it out for now.

Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Link: https://msgid.link/20240415112355.4d34b6a31be7.Ie8453979f5805873a8411c99346bcc3810cd6476@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>

K2CI-Arch: All
Change-Id: I4348a9e6c26c40c24d566a6adc02acc862387931
Signed-off-by: Yi Cong <yicong@kylinos.cn>
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/92073
Code-Style: kernel-bot <kernel-bot@kylinos.cn>
Tested-by: kernel-bot <kernel-bot@kylinos.cn>
Reviewed-by: 刘云 <liuyun01@kylinos.cn>