SUNWAY: mm: take memory information from firmware
Mainline: KYLIN-only
Severity: Low

Task: #382309

Currently, we directly detect the memory via I/O operations in kernel.
Considering that it is the responsibility of the firmware to mask
hardware differences, it is best for the firmware to detect memory
and pass the memory information to kernel via a standard interface.

The standard interface stated above mainly contains two aspects:

1. UEFI memory map

If UEFI BIOS is used to boot kernel, the UEFI BIOS is responsible for
detecting usable physical memory and passing it to kernel via UEFI
memory map.

2. Memory node in device tree

If UEFI BIOS is not used, the firmware in use is responsible for
detecting usable physical memory and passing it to kernel via memory
node in device tree.

An example of a memory node is as follows:

memory@0 {
	device_type = "memory";
	reg = <xxx>;        /* base address and size */
	numa-node-id = <0>; /* memory of NUMA node 0 */
}

This commit brings two changes:

1. Take memory information from firmware when receive the boot magic
"0xDEED2024".

2. Some refactoring, moving memblock related initialization from setup.c
to other files.

K2CI-Arch: Sunway
Signed-off-by: Jing Li <jingli@wxiat.com>
Signed-off-by: liuyunchen <liuyunchen@kylinos.cn>
Change-Id: I1f6b841ee35ec16cfc6352d9fd0265cbe91ba3cd
Reviewed-on: https://code.kylinos.cn:443/c/institute/kernel/klinux/+/63847
Code-Style: kernel-bot <kernel-bot@kylinos.cn>
Tested-by: kernel-bot <kernel-bot@kylinos.cn>
Reviewed-by: 刘云 <liuyun01@kylinos.cn>