KYLIN: i40e: revert "Partial revert "ethernet/intel: consolidate NAPI  and NAPI exit" "
Mainline: KYLIN-only
Severity: Low

Task: #211035

If the polling delay is set to busy_read and busy_poll,
the performance of the patch will be greatly improved. However,
the setting method is not the same as the performance optimization method.
Roll back the patch and align it with the performance optimization patches
in the polling mode of the community i40e. This patch has no effect.

K2CI-Arch: All
Signed-off-by: luoqing <luoqing@kylinos.cn>
Change-Id: Ib3840e93f9832b1b859475832269d16c40d498df
Reviewed-on: http://gerrit.kylin.com/c/klinux/+/107215
Code-Style: k2ci <kernel-bot@kylinos.cn>
Tested-by: k2ci <kernel-bot@kylinos.cn>
Reviewed-by: Jackie Liu <liuyun01@kylinos.cn>